FEAT: Move model geometries and BCs out of test_utils
