+++ src/porepy/applications/boundary_conditions/__init__.py
@@ -1,0 +1,1 @@
+from . import model_boundary_conditions
