+++ src/porepy/applications/test_utils/models.py
@@ -7,309 +7,10 @@ from typing import Any, Callable
 import numpy as np
 
 import porepy as pp
-from porepy.applications.md_grids.model_geometries import CubeDomainOrthogonalFractures
-
-
-class RectangularDomainThreeFractures(pp.ModelGeometry):
-    """A rectangular domain with up to three fractures.
-
-    The first two fractures are orthogonal, with `x` and `y` coordinates equal to
-    0.5, respectively. The third fracture is tilted. The number of fractures is
-    controlled by the parameter ``fracture_indices``, which can be any subset of
-    [0, 1, 2].
-
-    """
-
-    params: dict
-    """Parameters for the model."""
-
-    def set_fractures(self) -> None:
-        # Length scale:
-        ls = self.solid.convert_units(1, "m")
-
-        fracture_indices = self.params.get("fracture_indices", [0])
-        fractures = [
-            pp.LineFracture(np.array([[0, 2], [0.5, 0.5]]) * ls),
-            pp.LineFracture(np.array([[0.5, 0.5], [0, 1]]) * ls),
-            pp.LineFracture(np.array([[0.3, 0.7], [0.3, 0.7]]) * ls),
-        ]
-        self._fractures = [fractures[i] for i in fracture_indices]
-
-    def meshing_arguments(self) -> dict:
-        # Divide by length scale:
-        ls = self.solid.convert_units(1, "m")
-
-        mesh_sizes = {
-            # Cartesian: 2 by 8 cells.
-            "cell_size_x": 0.25 * ls,
-            "cell_size_y": 0.5 * ls,
-            # Simplex. Whatever gmsh decides.
-            "cell_size_fracture": 0.5 * ls,
-            "cell_size_boundary": 0.5 * ls,
-            "cell_size_min": 0.2 * ls,
-        }
-        return mesh_sizes
-
-    def set_domain(self) -> None:
-        if not self.params.get("cartesian", False):
-            self.params["grid_type"] = "simplex"
-        else:
-            self.params["grid_type"] = "cartesian"
-
-        # Length scale:
-        ls = self.solid.convert_units(1, "m")
-
-        # Mono-dimensional grid by default
-        phys_dims = np.array([2, 1]) * ls
-        box = {"xmin": 0, "xmax": phys_dims[0], "ymin": 0, "ymax": phys_dims[1]}
-        self._domain = pp.Domain(box)
-
-
-class OrthogonalFractures3d(CubeDomainOrthogonalFractures):
-    """A 3d domain with up to three orthogonal fractures.
-
-    The fractures have constant `x`, `y` and `z` coordinates equal to 0.5, respectively,
-    and are situated in a unit cube domain. The number of fractures is controlled by
-    the parameter ``num_fracs``, which can be 0, 1, 2 or 3.
-
-    """
-
-    params: dict
-    """Model parameters."""
-
-    def meshing_arguments(self) -> dict:
-        # Length scale:
-        ls = self.solid.convert_units(1, "m")
-
-        mesh_sizes = {
-            "cell_size": 0.5 * ls,
-            "cell_size_fracture": 0.5 * ls,
-            "cell_size_boundary": 0.5 * ls,
-            "cell_size_min": 0.2 * ls,
-        }
-        return mesh_sizes
-
-
-class WellGeometryMixin:
-    """Mixin class for models with wells."""
-
-    nd: int
-    """Number of dimensions."""
-    params: dict
-    """Model parameters."""
-
-    def set_well_network(self) -> None:
-        """Assign well network class."""
-        num_wells = self.params.get("num_wells", 1)
-        wells = [
-            # Intersects one (horizontal) fracture of OrthogonalFractures3d and
-            # extends to the top of the domain
-            pp.Well(np.array([[0.2, 0.2], [0.1, 0.1], [0.2, 1]])),
-            # Intersects two (horizontal and vertical) fractures of
-            # OrthogonalFractures3d. Extends between two domain boundaries.
-            pp.Well(np.array([[0.0, 0.6], [0.5, 0.5], [0.4, 1]])),
-            # Intersects no fractures. Internal well.
-            pp.Well(np.array([[0.3, 0.3], [0.3, 0.3], [0.3, 0.4]])),
-        ]
-        parameters = {"mesh_size": 0.2}
-        self.well_network = pp.WellNetwork3d(
-            self.domain, wells[:num_wells], parameters=parameters
-        )
-
-
-class BoundaryConditionsMassDirNorthSouth:
-    """Boundary conditions for the flow problem.
-
-    Dirichlet boundary conditions are defined on the north and south boundaries. Some
-    of the default values may be changed directly through attributes of the class.
-
-    Usage: tests for models defining equations for any subset of the thermoporomechanics
-    problem.
-
-    """
-
-    def bc_type_darcy_flux(self, sd: pp.Grid) -> pp.BoundaryCondition:
-        """Boundary condition type for Darcy flux.
-
-        Dirichlet boundary conditions are defined on the north and south boundaries.
-
-        Parameters:
-            sd: Subdomain for which to define boundary conditions.
-
-        Returns:
-            bc: Boundary condition object.
-
-        """
-        domain_sides = self.domain_boundary_sides(sd)
-        # Define boundary condition on faces
-        return pp.BoundaryCondition(sd, domain_sides.north + domain_sides.south, "dir")
-
-    def bc_values_pressure(self, boundary_grid: pp.BoundaryGrid) -> np.ndarray:
-        """Boundary condition values for Darcy flux.
-
-        Dirichlet boundary conditions are defined on the north and south boundaries,
-        with a constant value of 0 unless fluid's reference pressure is changed.
-
-        Parameters:
-            boundary_grid: Boundary grid for which to define boundary conditions.
-
-        Returns:
-            Boundary condition values array.
-
-        """
-        domain_sides = self.domain_boundary_sides(boundary_grid)
-        vals_loc = np.zeros(boundary_grid.num_cells)
-        vals_loc[domain_sides.north + domain_sides.south] = self.fluid.pressure()
-        return vals_loc
-
-    def bc_type_fluid_flux(self, sd):
-        """Boundary condition type for the density-mobility product.
-
-        Dirichlet boundary conditions are defined on the north and south boundaries.
-
-        Parameters:
-            sd: Subdomain for which to define boundary conditions.
-
-        Returns:
-            bc: Boundary condition object.
-
-        """
-        domain_sides = self.domain_boundary_sides(sd)
-        # Define boundary condition on faces
-        return pp.BoundaryCondition(sd, domain_sides.north + domain_sides.south, "dir")
-
-
-class BoundaryConditionsEnergyDirNorthSouth(pp.BoundaryConditionMixin):
-    """Boundary conditions for the thermal problem.
-
-    Dirichlet boundary conditions are defined on the north and south boundaries. Some
-    of the default values may be changed directly through attributes of the class.
-
-    Usage: tests for models defining equations for any subset of the thermoporomechanics
-    problem.
-
-    """
-
-    def bc_type_fourier_flux(self, sd: pp.Grid) -> pp.BoundaryCondition:
-        """Boundary condition type for the Fourier heat flux.
-
-        Dirichlet boundary conditions are defined on the north and south boundaries.
-
-        Parameters:
-            sd: Subdomain for which to define boundary conditions.
-
-        Returns:
-            bc: Boundary condition object.
-
-        """
-        domain_sides = self.domain_boundary_sides(sd)
-        # Define boundary condition on faces
-        return pp.BoundaryCondition(sd, domain_sides.north + domain_sides.south, "dir")
-
-    def bc_type_enthalpy_flux(self, sd: pp.Grid) -> pp.BoundaryCondition:
-        """Boundary condition type for the enthalpy.
-
-        Dirichlet boundary conditions are defined on the north and south boundaries.
-
-        Parameters:
-            sd: Subdomain for which to define boundary conditions.
-
-        Returns:
-            bc: Boundary condition object.
-
-        """
-        domain_sides = self.domain_boundary_sides(sd)
-        # Define boundary condition on faces
-        return pp.BoundaryCondition(sd, domain_sides.north + domain_sides.south, "dir")
-
-
-class BoundaryConditionsMechanicsDirNorthSouth(pp.BoundaryConditionMixin):
-    """Boundary conditions for the mechanics with Dirichlet conditions on north and
-    south boundaries.
-
-    """
-
-    params: dict[str, Any]
-    """Model parameters."""
-    solid: pp.SolidConstants
-    """Solid parameters."""
-    fluid: pp.FluidConstants
-    """Fluid parameters."""
-
-    def bc_type_mechanics(self, sd):
-        """Boundary condition type for mechanics.
-
-        Dirichlet boundary conditions are defined on the north and south boundaries.
-
-        Parameters:
-            sd: Subdomain for which to define boundary conditions.
-
-        Returns:
-            bc: Boundary condition object.
-
-        """
-        domain_sides = self.domain_boundary_sides(sd)
-        bc = pp.BoundaryConditionVectorial(
-            sd, domain_sides.north + domain_sides.south, "dir"
-        )
-        bc.internal_to_dirichlet(sd)
-        return bc
-
-    def bc_values_displacement(self, boundary_grid: pp.BoundaryGrid) -> np.ndarray:
-        """Boundary values for the mechanics problem as a numpy array.
-
-        Extracted from below method to facilitate time dependent boundary conditions.
-        Values for north and south faces are set to zero unless otherwise specified
-        through attributes ux_north, uy_north, ux_south, uy_south.
-
-        Parameters:
-            boundary_grid: Boundary grid for which boundary values are to be returned.
-
-        Returns:
-            Array of boundary values, with one value for each dimension of the
-                problem, for each face in the subdomain.
-
-        """
-        domain_sides = self.domain_boundary_sides(boundary_grid)
-        values = np.zeros((self.nd, boundary_grid.num_cells))
-        values[1, domain_sides.north] = self.solid.convert_units(
-            self.params.get("uy_north", 0), "m"
-        )
-        values[1, domain_sides.south] = self.solid.convert_units(
-            self.params.get("uy_south", 0), "m"
-        )
-        values[0, domain_sides.north] = self.solid.convert_units(
-            self.params.get("ux_north", 0), "m"
-        )
-        values[0, domain_sides.south] = self.solid.convert_units(
-            self.params.get("ux_south", 0), "m"
-        )
-        return values.ravel("F")
-
-
-class TimeDependentMechanicalBCsDirNorthSouth(BoundaryConditionsMechanicsDirNorthSouth):
-    """Time dependent displacement boundary conditions.
-
-    For use in (thermo)poremechanics.
-    """
-
-    def bc_values_displacement(self, boundary_grid: pp.BoundaryGrid) -> np.ndarray:
-        domain_sides = self.domain_boundary_sides(boundary_grid)
-        values = np.zeros((self.nd, boundary_grid.num_cells))
-        # Add fracture width on top if there is a fracture.
-        if len(self.mdg.subdomains()) > 1:
-            frac_val = self.solid.convert_units(0.042, "m")
-        else:
-            frac_val = 0
-        values[1, domain_sides.north] = frac_val
-        if self.time_manager.time > 1e-5:
-            values[1, domain_sides.north] += self.solid.convert_units(
-                self.params.get("uy_north", 0), "m"
-            )
-            values[1, domain_sides.south] += self.solid.convert_units(
-                self.params.get("uy_south", 0), "m"
-            )
-        return values.ravel("F")
+from porepy.applications.md_grids.model_geometries import (
+    RectangularDomainThreeFractures,
+    OrthogonalFractures3d,
+)
 
 
 class NoPhysics(
@@ -471,7 +172,7 @@ def _add_mixin(mixin, parent):
     return cls
 
 
-# TODO: Move. Check all values.
+# TODO: Purge in favour of pp.solid_values.granite
 granite_values = {
     "biot_coefficient": 0.8,
     "permeability": 1e-20,
@@ -487,7 +188,6 @@ granite_values = {
     "fracture_gap": 1e-4,
     "residual_aperture": 0.01,
 }
-# Cf. fluid.py
 water_values = {
     "specific_heat_capacity": 4180,
     "compressibility": 4e-10,
