+++ tests/models/test_momentum_balance.py
@@ -11,14 +11,16 @@ import pytest
 import porepy as pp
 
 from porepy.applications.test_utils.models import (
-    BoundaryConditionsMechanicsDirNorthSouth,
     MomentumBalance,
     compare_scaled_model_quantities,
     compare_scaled_primary_variables,
 )
 
 
-class LinearModel(BoundaryConditionsMechanicsDirNorthSouth, MomentumBalance):
+class LinearModel(
+    pp.model_boundary_conditions.BoundaryConditionsMechanicsDirNorthSouth,
+    MomentumBalance,
+):
     pass
