+++ tests/models/test_solution_strategy.py
@@ -220,8 +220,8 @@ class RediscretizationTest:
 
 # Non-trivial solution achieved through BCs.
 class RandomPressureBCs(
-    models.BoundaryConditionsMassDirNorthSouth,
-    models.BoundaryConditionsEnergyDirNorthSouth,
+    pp.model_boundary_conditions.BoundaryConditionsMassDirNorthSouth,
+    pp.model_boundary_conditions.BoundaryConditionsEnergyDirNorthSouth,
 ):
     def bc_values_pressure(self, boundary_grid: pp.BoundaryGrid) -> np.ndarray:
         """Boundary condition values for Darcy flux.
